% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree.R
\name{.get_tree}
\alias{.get_tree}
\title{Get the ontology tree}
\usage{
.get_tree(concepts, topID)
}
\arguments{
\item{concepts}{\code{\link[=tibble]{tibble(1)}}\cr the concepts table of an ontology.}

\item{topID}{[][]\cr the ID from which to derive the tree}
}
\description{
Get the ontology tree
}
