% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTreeMap.R
\name{.makeTreeMap}
\alias{.makeTreeMap}
\title{Make a tree map}
\usage{
.makeTreeMap(onto = NULL)
}
\arguments{
\item{onto}{\code{\link[=list]{ontology(1)}}\cr an already loaded ontology.}
}
\description{
A tree map is used via the show method of an onto object.
}
