% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_matches.R
\name{edit_matches}
\alias{edit_matches}
\title{Edit matches manually in a csv-table}
\usage{
edit_matches(
  concepts,
  attributes = NULL,
  source = NULL,
  ontology = NULL,
  matchDir = NULL,
  verbose = TRUE
)
}
\arguments{
\item{concepts}{\code{\link[=data.frame]{data.frame(.)}}\cr the new concepts that shall
be manually matched.}

\item{attributes}{\code{\link[=character]{character(1)}}\cr the attributes of new
concepts that help to match new and target concepts concepts manually (must
contain at least the column 'class').}

\item{source}{\code{\link[=character]{character(1)}}\cr any character uniquely
identifying the source dataset of the new concepts.}

\item{ontology}{\code{\link[=list]{ontology(1)}}\cr either a path where the ontology is
stored, or an already loaded ontology.}

\item{matchDir}{\code{\link[=character]{character(1)}}\cr the directory where to store
source-specific matching tables.}

\item{verbose}{\code{\link[=logical]{logical(1)}}\cr whether or not to give detailed
information on the process of this function.}
}
\description{
Edit matches manually in a csv-table
}
