% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tree.R
\name{get_tree}
\alias{get_tree}
\title{Get the ontology tree}
\usage{
get_tree(onto, top)
}
\arguments{
\item{onto}{\code{\link[=tibble]{tibble(1)}}\cr the concepts table of an ontology.}

\item{top}{\code{\link[=character]{character(1)}}\cr the ID from which to derive the
tree.}
}
\description{
Get the ontology tree
}
