% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ontology.R
\name{load_ontology}
\alias{load_ontology}
\title{Load a gazetteer}
\usage{
load_ontology(ontoDir)
}
\arguments{
\item{ontoDir}{\code{\link[=character]{character(1)}}\cr the path where the ontology to
load is stored. It can be omitted in case the option "onto_path" has been
define (see \code{getOption("onto_path")}).}
}
\description{
Load a gazetteer
}
\details{
A \href{https://en.wikipedia.org/wiki/Gazetteer}{gezatteer} is a
directory "that contains information concerning the geographical makeup,
social statistics and physical features of a country, region, or
continent", or in the case here, additionally the ontological relations of
the areal data.
}
