% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_concept.R
\name{new_concept}
\alias{new_concept}
\title{Add a new concept to an ontology}
\usage{
new_concept(
  new,
  broader = NULL,
  class = NULL,
  source,
  attributes = NULL,
  ontology = NULL
)
}
\arguments{
\item{new}{\code{\link[=character]{character(.)}}\cr the english label(s) of new
concepts that shall be included in the ontology.}

\item{broader}{\code{\link[=data.frame]{data.frame(.)}}\cr the english label(s) of already
harmonised concepts to which the new concept shall be semantically linked
via a
\href{https://www.w3.org/TR/skos-reference/#semantic-relations}{skos:broader}
relation, see Details.}

\item{class}{\code{\link[=character]{character(.)}}\cr the class(es) of the new labels.}

\item{source}{\code{\link[=character]{character(1)}}\cr any character uniquely
identifying the source dataset of the new concept (for example
\emph{Author+Year}).}

\item{attributes}{\code{\link[=tibble]{tibble()}}\cr not yet implemented.}

\item{ontology}{\code{\link[=list]{ontology(1)}}\cr either a path where the
ontology is stored, or an already loaded ontology.}
}
\value{
returns invisibly a table of the new harmonised concepts that were
added to the ontology, or a message that nothing new was added.
}
\description{
This adds a new concept to an existing ontology to semantically integrate and
thus harmonise it with the already existing ontology.
}
\examples{
ontoDir <- system.file("extdata", "crops.rds", package = "ontologics")
onto <- load_ontology(path = ontoDir)

concepts <- data.frame(old = c("Bioenergy woody", "Bioenergy herbaceous"),
                       new = c("acacia", "miscanthus"))

onto <- new_source(name = "externalDataset",
                   description = "a vocabulary",
                   homepage = "https://www.something.net",
                   license = "CC-BY-0",
                   ontology = onto)

onto <- get_concept(terms = concepts$old, ontology = onto) \%>\%
  new_concept(new = concepts$new,
              broader = .,
              class = "crop",
              source = "externalDataset",
              ontology = onto)
}
