% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_mapping.R
\name{new_mapping}
\alias{new_mapping}
\title{Add a new mapping to an ontology}
\usage{
new_mapping(
  new = NULL,
  concept,
  match = "close",
  source = NULL,
  mappings = NULL,
  certainty = NULL,
  ontology = NULL
)
}
\arguments{
\item{new}{\code{\link[=character]{character(.)}}\cr the english label(s) of new
external concepts that shall be mapped to concepts that do already exist in
the ontology.}

\item{concept}{\code{\link[=data.frame]{data.frame(.)}}\cr the english label(s) of
already harmonised concepts to which the external concepts shall be mapped.}

\item{match}{\code{\link[=character]{character(1)}}\cr the
\href{https://www.w3.org/TR/skos-reference/#mapping}{skos mapping property}
used to describe the link, possible values are \code{"close"},
\code{"exact"}, \code{"broad"}, \code{"narrow"} and \code{"related"}.}

\item{source}{\code{\link[=character]{character(1)}}\cr any character uniquely
identifying the source dataset of the new concept.}

\item{mappings}{\code{\link[=tibble]{tibble()}}\cr not yet implemented.}

\item{certainty}{\code{\link[=integer]{integerish(1)}}\cr the certainty of the match.
Possible values are between 1 and 4, with meaning \itemize{\item 1 =
probably unreliable \item 2 = unclear, assigned according to a given
definition \item 3 = clear, assigned according to a given definition \item
4 = original, harmonised term (can't be assigned by a user)}.}

\item{ontology}{\code{\link[=list]{ontology(1)}}\cr either a path where the ontology is
stored, or an already loaded ontology.}
}
\value{
No return value, called for the side effect of adding new mappings to
an ontology.
}
\description{
Extend an ontology by creating mappings to external vocabularies.
}
\examples{
ontoDir <- system.file("extdata", "crops.rds", package = "ontologics")
onto <- load_ontology(path = ontoDir)

mapping <- data.frame(old = c("BIOENERGY CROPS", "Bioenergy woody",
                              "Other bioenergy crops"),
                      new = c("bioenergy plants", "Wood plantation for fuel",
                              "Algae for bioenergy"),
                      type = c("close", "broad", "broad"))

onto <- new_source(name = "externalDataset",
                   description = "a vocabulary",
                   homepage = "https://www.something.net",
                   license = "CC-BY-0",
                   ontology = onto)

onto <- get_concept(x = data.frame(label_en = mapping$old), ontology = onto) \%>\%
  new_mapping(new = mapping$new,
              concept = .,
              match = mapping$type,
              source = "externalDataset",
              certainty = 3,
              ontology = onto)

}
