% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_source.R
\name{new_source}
\alias{new_source}
\title{Add a new valid source to an ontology}
\usage{
new_source(
  name = NULL,
  description = NULL,
  homepage = NULL,
  license = NULL,
  notes = NULL,
  ontology = NULL
)
}
\arguments{
\item{name}{\code{\link[=character]{character(1)}}\cr the name of the new source (should
not contain empty spaces).}

\item{description}{\code{\link[=character]{character(1)}}\cr a verbatim description of
the new source.}

\item{homepage}{\code{\link[=character]{character(1)}}\cr the homepage of the new source
(this could be the basis of the concept URIs).}

\item{license}{\code{\link[=character]{character(1)}}\cr the licenses under which the
new source is published.}

\item{notes}{\code{\link[=character]{character(1)}}\cr any notes on the new source that
don't fit into any of the other meta-data fields here.}

\item{ontology}{\code{\link[=list]{ontology(1)}}\cr either a path where the ontology is
stored, or an already loaded ontology.}
}
\description{
Add a new valid source to an ontology
}
\examples{
ontoDir <- system.file("extdata", "crops.rds", package = "ontologics")
onto <- load_ontology(path = ontoDir)

onto <- new_source(name = "externalDataset",
                   description = "a vocabulary",
                   homepage = "https://www.something.net",
                   license = "CC-BY-0",
                   ontology = onto)
}
