% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newConcept.R
\name{newConcept}
\alias{newConcept}
\title{Add a new concept to an ontology}
\usage{
newConcept(
  new,
  broader,
  class = NULL,
  source,
  only_new = TRUE,
  attributes = NULL,
  ontoDir = NULL
)
}
\arguments{
\item{new}{\code{\link[=character]{character(.)}}\cr the english label(s) of new
concepts that shall be included in the ontology.}

\item{broader}{\code{\link[=character]{character(.)}}\cr the english label(s) of already
harmonised concepts to which the new concept shall be semantically linked
via a
\href{https://www.w3.org/TR/skos-reference/#semantic-relations}{skos:broader}
relation.}

\item{class}{\code{\link[=character]{character(.)}}\cr the class(es) to of the new
labels.}

\item{source}{\code{\link[=character]{character(1)}}\cr any character uniquely
identifying the source dataset of the new concept.}

\item{only_new}{\code{\link[=logical]{logical(1)}} ?}

\item{attributes}{\code{\link[=tibble]{tibble()}}\cr not yet implemented.}

\item{ontoDir}{\code{\link[=character]{character(1)}}\cr the path where the ontology to
update is stored. In case you are building a point or areal database and
have used the function \code{\link[luckiTools]{start_occurrenceDB}} or
\code{\link[luckiTools]{start_arealDB}}, this path is already stored in the
options (see \code{getOption("onto_path")}).}
}
\description{
Add a new concept to an ontology
}
\seealso{
Other ontology functions: 
\code{\link{getConcept}()},
\code{\link{newMapping}()}
}
\concept{ontology functions}
