% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1onto.R
\name{show,onto-method}
\alias{show,onto-method}
\title{Print onto in the console}
\usage{
\S4method{show}{onto}(object)
}
\arguments{
\item{object}{object to \code{show}.}
}
\description{
Print onto in the console
}
