% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_ontology.R
\name{start_ontology}
\alias{start_ontology}
\title{Start an ontology}
\usage{
start_ontology(
  name = NULL,
  path = NULL,
  code = ".xx",
  description = NULL,
  homepage = NULL,
  license = NULL,
  notes = NULL
)
}
\arguments{
\item{name}{\code{\link[=character]{character(1)}}\cr the path of the ontology.}

\item{path}{\code{\link[=character]{character(1)}}\cr the path where the ontology shall
be stored.}

\item{code}{\code{\link[=double]{double(1)}}\cr format of a single code snippet that is
concatenated for nested levels.}

\item{description}{\code{\link[=character]{character(1)}}\cr}

\item{homepage}{\code{\link[=character]{character(1)}}\cr}

\item{license}{\code{\link[=character]{character(1)}}\cr}

\item{notes}{\code{\link[=character]{character(1)}}\cr}
}
\value{
it returns the new, emtpy ontology and also stores that within the
directory specified in \code{path}.
}
\description{
Start an ontology
}
\examples{
start_ontology(name = "crops", path = tempdir())
}
