% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tree.R
\name{make_tree}
\alias{make_tree}
\title{Make a tree of an ontology}
\usage{
make_tree(top, ontology = NULL)
}
\arguments{
\item{top}{\code{\link[=tibble]{tibble(1)}}\cr the concepts table that shall be at the
top of the tree.}

\item{ontology}{\code{\link[=list]{ontology(1)}}\cr either a path where the ontology is
stored, or an already loaded ontology.}
}
\description{
Make a tree of an ontology
}
