% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tree.R
\name{make_tree}
\alias{make_tree}
\title{Make a tree of an ontology}
\usage{
make_tree(input, top)
}
\arguments{
\item{input}{\code{\link[=tibble]{tibble(1)}}\cr the concepts table of an ontology.}

\item{top}{\code{\link[=character]{character(1)}}\cr the ID from which to derive the
tree.}
}
\description{
Make a tree of an ontology
}
