% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_class.R
\name{new_class}
\alias{new_class}
\title{Add a new valid class to an ontology}
\usage{
new_class(class, parent, ontology = NULL)
}
\arguments{
\item{class}{\code{\link[=character]{character(1)}}\cr}

\item{parent}{\code{\link[=character]{character(1)}}\cr}

\item{ontology}{\code{\link[=list]{ontology(1)}}\cr either a path where the
ontology is stored, or an already loaded ontology.}
}
\description{
Add a new valid class to an ontology
}
\examples{
ontoDir <- system.file("extdata", "crops.rds", package = "ontologics")
onto <- load_ontology(path = ontoDir)

onto <- new_class(class = "group", parent = NA, ontology = onto) \%>\%
        new_class(class = "class", parent = "group", ontology = .) \%>\%
        new_class(class = "crop", parent = "class", ontology = .)

}
