% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_class.R
\name{new_class}
\alias{new_class}
\title{Add a new valid class to an ontology}
\usage{
new_class(class, broader, description, ontology = NULL)
}
\arguments{
\item{class}{\code{\link[=character]{character(1)}}\cr the new class label.}

\item{broader}{\code{\link[=character]{character(1)}}\cr the broader class into which
the new class is nested.}

\item{description}{\code{\link[=character]{character(1)}}\cr a verbatim description of
the new class.}

\item{ontology}{\code{\link[=list]{ontology(1)}}\cr either a path where the ontology is
stored, or an already loaded ontology.}
}
\description{
Add a new valid class to an ontology
}
\examples{
ontoDir <- system.file("extdata", "crops.rds", package = "ontologics")
onto <- load_ontology(path = ontoDir)

onto <- new_class(class = "use type", broader = "class",
                  description = "something", ontology = onto)

}
