% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1onto.R
\docType{class}
\name{onto-class}
\alias{onto-class}
\alias{onto}
\title{Ontology class (S4) and methods}
\description{
Ontology class (S4) and methods
}
\section{Slots}{

\describe{
\item{\code{sources}}{\code{\link[=data.frame]{data.frame(.)}}\cr}

\item{\code{classes}}{\code{\link[=data.frame]{data.frame(.)}}\cr}

\item{\code{concepts}}{\code{\link[=data.frame]{data.frame(.)}}\cr}
}}

