% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_ontology.R
\name{start_ontology}
\alias{start_ontology}
\title{Start an ontology}
\usage{
start_ontology(
  name = NULL,
  version = NULL,
  path = NULL,
  code = ".xx",
  description = NULL,
  homepage = NULL,
  uri_prefix = NULL,
  license = NULL,
  notes = NULL
)
}
\arguments{
\item{name}{\code{\link[=character]{character(1)}}\cr the path of the ontology.}

\item{version}{\code{\link[=character]{character(1)}}\cr the version of the ontology.}

\item{path}{\code{\link[=character]{character(1)}}\cr the path where the ontology shall
be stored.}

\item{code}{\code{\link[=double]{double(1)}}\cr format of a single code snippet that is
concatenated for nested levels.}

\item{description}{\code{\link[=character]{character(1)}}\cr a brief description of the
new ontology.}

\item{homepage}{\code{\link[=character]{character(1)}}\cr the URL to the homepage of the
new ontology.}

\item{uri_prefix}{\code{\link[=character]{character(1)}}\cr the basic URL to construct
URIs for all concepts.}

\item{license}{\code{\link[=character]{character(1)}}\cr any string describing the
license under which this ontology can be (re)used.}

\item{notes}{\code{\link[=character]{character(1)}}\cr any notes that might apply to
this ontology.}
}
\value{
it returns the new, empty ontology and also stores that within the
directory specified in \code{path}.
}
\description{
Start an ontology
}
\examples{
start_ontology(name = "crops", path = tempdir())
}
